backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
sendgrid==6.10.0
slack-sdk==3.24.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0
